task: Integrate with frontend
Frontend tiled using the new css property aspect-ratio. Uncertain of the
support for this, but it's a much simpler tiling mechanism than the
`padding-top` hack I used for griddle.

Implemented `Letter` class to abstract some of the Javascript logic
from the HTML elements. Each `Letter` object maps to one of the buttons
on the bottom row to avoid having to interact with the HTML directly
from within the Javascript.

The user selection is rendered at the top by simply selecting the next
"unselected" element in the row. Eventually, the user will be able to
clear the selection or submit the selection. If the submission is a
word, it will get rendered in the `output-container` div.

The output-container div currently goes behind the buttons at the
bottom. Some padding will be required to ensure the last words in the
grid are still visible.