feat: Enable entering input
Validates the selected word. If the word is valid, it's added to the
output screen. If the word is invalid, a log message is displayed for
now.

To ensure all found words are visible, a few dud `letter-containers` are
added to the `output-container`. This pads the end of the container,
such that no "real" words are ever hidden.

The output containers are rendering using Javascript template strings.
I'm usually hesitant to use these for XSS reasons, but it should be safe
here as all of the content is generated and displayed within the local
webpage.