refactor: Reduce letter count
In an effort to make the game more mobile-friendly, I've reduced the
number of letter squares from 10 to 8.

I think 10 makes for a more interesting game, but it's visually a lot of
infromation to process and the squares are very hard to tap on a small
screen.

The other alternative is to do two rows of 5 squares, but then the
output doesn't line up nicely with the input letters.

The end goal is to generate all of the possible words anyway, so fewer
letters = fewer words to compute. It's not a compromise, it's a
performance enhancement!