refactor(input): Move `selection-container` to the bottom
I found myself glancing up at the output repeatedly while inputting
letters, particularly on mobile. I think this is because the input
letters are obviously non-standard vs a qwerty keyboard, so one has to
look at both the input and the output letters simultaneously to ensure
the correct word is being entered.

By placing the output above the input, the UX of entering a word should
be improved.

Incidentally, the CSS is clearly working well because this was a super
trivial move which aligned nicely with the existing layout.