refactor: Shift the colour palette to green
The game is looking a little grey, so this should make it visually less
boring. The selected selection letters are already green, so I'm leaning
towards green.

Tried to make the output letters sort of "inactive" but at the same time
visible.

Also made the border and text colour off-white to reduce the sharp
contrast with the background.