feat: Add shuffle button
Shuffles the letters randomly to make it easier to find new words if the
game is becoming stale.

Removed the button words for scaling reasons.

Introduced a fourth button for quitting the game, but unsure whether to
use that for something else yet. Depends how the game develops. For now,
this just keeps the game proportions looking right.