refactor(output-container): Add explicit output word container
Currently if we clear the `output-container`, the padding (and any other
content) will also be cleared.

When we clear the output container, we only want the output words to be
cleared, no the padding.

This ensures the output words are added to their own dedicated
container. If we clear this container, only the output words are
cleared. The padding is retained.