feat: Add new-game button
This will reset the game to the starting state, with a new series of
input letters.

This is obviously a dangerous action, so I've made the reset button
explciitly dangerous (red). To compliment this style, I've made the
"submit" button green.

I've also changed the "clear" button icon from a bin to an "undo" icon
to differentiate its purpose from the "new game" button.