refactor(is_word): Lookup words in dict object
In Javascript, [dict key lookup is O(1)][1], which is more efficient
than `Array.includes`.

`Set` may also be a good option. I think the time complexity is slightly
worse, _but_ `WORDLIST` currently has a bunch of unused values to its
keys, which will increase the size of `wordlist.js` and therefore the
load time of the page.

[1]: https://stackoverflow.com/a/23500629