task: Generate all valid words
This is required to determine how many of the possible words that the
player has found. For now this is a pretty simple recursive function,
but it could be improved by bucketing the words by length. From this, we
could give the player some sense of which words they haven't yet found.