task: Increase `MIN_WORD_LENGTH`
I find that the 2-letter words are mostly nonsense, stuff like `oe` that
a normal human being wouldn't think to play. It's also a bit too easy
for other normal 2-letter words, and bloats the number of words that a
player has to find. As such, this has been increased to `3` - and could
probably do with being increased further, because 3-letter words are not
much fun either!